OpenMPTarget: DeepCopy in separate file.
